Fix knitr caching when same process is used
If the same process is used for all chunks, we
cannot save the process object in the knitr options,
because the options are part of the hash for the caching,
and the process will be different every time.

Ideally, R objects would have a cache method to be
used for this, but of course that does not exist.

So we just put the process in the global env.